remove old database files, rename db to database
