+++ source/utils/database/errors.py
