+++ source/utils/database/requests/global_.py
