--- source/utils/db/__init__.py
@@ -1,3 +1,0 @@
-from .client import DataBaseClient
-from .models import *  # noqa: F401 F403
-from .requests import *  # noqa: F401 F403
